feat(server): Booking and User * Add more mock data to db seed script & add seed clear script * Defined Database document structure

+++ server/src/graphql/typeDefs.ts
@@ -1,4 +1,4 @@
-import { gql } from 'apollo-server-express';
+import { gql } from "apollo-server-express";
 
 export const typeDefs = gql`
   type Listing {
@@ -10,13 +10,11 @@ export const typeDefs = gql`
     numOfGuests: Int!
     numOfBeds: Int!
     numOfBaths: Int!
-    rating: Float!
+    rating: Int!
   }
-
   type Query {
     listings: [Listing!]!
   }
-
   type Mutation {
     deleteListing(id: ID!): Listing!
   }
